+++ Mediapipe.Net.Tests/Framework/Packets/BoolPacketTest.cs
@@ -3,20 +3,29 @@
 // MediaPipe.NET is licensed under the MIT License. See LICENSE for details.
 
 using Mediapipe.Net.Framework;
-using Mediapipe.Net.Framework.Packet;
+using Mediapipe.Net.Framework.Packets;
 using NUnit.Framework;
 
 namespace Mediapipe.Net.Tests.Framework.NewPacket
 {
-    public class FloatPacketTest
+    public class BoolPacketTest
     {
         #region Constructor
         [Test]
-        public void Ctor_ShouldInstantiatePacket_When_CalledWithValue()
+        public void Ctor_ShouldInstantiatePacket_When_CalledWithTrue()
         {
-            using var packet = PacketFactory.FloatPacket(0.01f);
-            Assert.True(packet.ValidateAsFloat().Ok());
-            Assert.AreEqual(packet.GetFloat(), 0.01f);
+            using var packet = PacketFactory.BoolPacket(true);
+            Assert.True(packet.ValidateAsBool().Ok());
+            Assert.True(packet.GetBool());
+            Assert.AreEqual(packet.Timestamp(), Timestamp.Unset());
+        }
+
+        [Test]
+        public void Ctor_ShouldInstantiatePacket_When_CalledWithFalse()
+        {
+            using var packet = PacketFactory.BoolPacket(false);
+            Assert.True(packet.ValidateAsBool().Ok());
+            Assert.False(packet.GetBool());
             Assert.AreEqual(packet.Timestamp(), Timestamp.Unset());
         }
 
@@ -24,9 +33,9 @@ namespace Mediapipe.Net.Tests.Framework.NewPacket
         public void Ctor_ShouldInstantiatePacket_When_CalledWithValueAndTimestamp()
         {
             using var timestamp = new Timestamp(1);
-            using var packet = PacketFactory.FloatPacket(0.01f).At(timestamp);
-            Assert.True(packet.ValidateAsFloat().Ok());
-            Assert.AreEqual(packet.GetFloat(), 0.01f);
+            using var packet = PacketFactory.BoolPacket(true).At(timestamp);
+            Assert.True(packet.ValidateAsBool().Ok());
+            Assert.True(packet.GetBool());
             Assert.AreEqual(packet.Timestamp(), timestamp);
         }
         #endregion
@@ -35,14 +44,14 @@ namespace Mediapipe.Net.Tests.Framework.NewPacket
         [Test]
         public void IsDisposed_ShouldReturnFalse_When_NotDisposedYet()
         {
-            using var packet = PacketFactory.FloatPacket(0f);
+            using var packet = PacketFactory.BoolPacket(true);
             Assert.False(packet.IsDisposed);
         }
 
         [Test]
         public void IsDisposed_ShouldReturnTrue_When_AlreadyDisposed()
         {
-            var packet = PacketFactory.FloatPacket(0f);
+            var packet = PacketFactory.BoolPacket(true);
             packet.Dispose();
 
             Assert.True(packet.IsDisposed);
@@ -51,10 +60,10 @@ namespace Mediapipe.Net.Tests.Framework.NewPacket
 
         #region DebugTypeName
         [Test]
-        public void DebugTypeName_ShouldReturnFloat_When_ValueIsSet()
+        public void DebugTypeName_ShouldReturnBool_When_ValueIsSet()
         {
-            using var packet = PacketFactory.FloatPacket(0.01f);
-            Assert.AreEqual(packet.DebugTypeName(), "float");
+            using var packet = PacketFactory.BoolPacket(true);
+            Assert.AreEqual(packet.DebugTypeName(), "bool");
         }
         #endregion
     }
