+++ Mediapipe.Net/Framework/Packets/Packet.cs
@@ -9,7 +9,7 @@ using Mediapipe.Net.Framework.Port;
 using Mediapipe.Net.Gpu;
 using Mediapipe.Net.Native;
 
-namespace Mediapipe.Net.Framework.Packet
+namespace Mediapipe.Net.Framework.Packets
 {
     public unsafe partial class Packet : MpResourceHandle
     {
