+++ Mediapipe.Net/Framework/Packets/PacketFactory.cs
@@ -8,7 +8,7 @@ using Mediapipe.Net.Gpu;
 using Mediapipe.Net.Graphs.InstantMotionTracking;
 using Mediapipe.Net.Native;
 
-namespace Mediapipe.Net.Framework.Packet
+namespace Mediapipe.Net.Framework.Packets
 {
     public static unsafe class PacketFactory
     {
