+++ Mediapipe.Net/Framework/Packets/PacketGetters.cs
@@ -12,7 +12,7 @@ using Mediapipe.Net.Graphs.InstantMotionTracking;
 using Mediapipe.Net.Native;
 using Mediapipe.Net.Util;
 
-namespace Mediapipe.Net.Framework.Packet
+namespace Mediapipe.Net.Framework.Packets
 {
     public unsafe partial class Packet : MpResourceHandle
     {
